Links: new feature allowing to provide unexecuted code in a link option. [[ my title | mylink | my {{ "code" }} ]] Must be placed in third position. The code won't be executed and will be kept as a string in the PassageLink of the PassageResult. This allow the user to have the possiblity of running code in between passages, depending of the link that the player will chose. It can be used to update variables for example, or display a short bit of text. It can even contain its own links, but I would recommend to keep things simple.
I had to remove a little feature to make this work: links like using the arrow, [[ title -> destination ]] or [[ destination <- title ]], is now not possible.
It did make sense to have a direction arrow with only 2 options, but with the introduction of a third option, it didn't make sense anymore. So I simply removed it.